Eliminate `TestEventContextFactory`
